+++ packages/regal/lib/src/widgets/widgets.dart
@@ -13,6 +13,8 @@ export 'infinite_listview.dart';
 export 'ledger_bottom_sheet.dart';
 export 'ledger_button.dart';
 export 'ledger_checkbox.dart';
+export 'ledger_date_picker.dart';
+export 'ledger_infinite_spinner.dart';
 export 'ledger_text_field.dart';
 export 'progress_timeline.dart';
 export 'regal_app_bar.dart';